feat(daily-update-r2): Wave 3 (6 parallel agents) — P2 hoist Group A (011/012/013) + auth migration (054) + P2b tests (032) + BFF size (042, +0.00 MB)
011 (P2 components hoist): all 9 FR-04 components hoisted to @spaarke/daily-briefing-components/src/components/:
DailyBriefingApp (renamed from App; shim exports both), TldrSection, ActivityNotesSection, ChannelHeading,
NarrativeBullet, PreferencesDropdown, CaughtUpFooter, DigestHeader, EmptyState + channelIcons helper. Original
locations under src/solutions/DailyBriefing/src/components/ converted to re-export shims pointing at
@spaarke/daily-briefing-components/components. NarrativeBullet logic preserved verbatim (P2a sub-list is task 020).
Cross-package alias resolution failures expected — task 016 wires path aliases.

012 (P2 briefingService hoist): briefingService.ts moved to @spaarke/daily-briefing-components/src/services/.
All exports (fetchAiBriefing, fetchBriefingNarration + types) re-exported from ./services subpath. Original
location is thin re-export shim. Logic byte-equivalent. Imports adjusted: authenticatedFetch via relative path
to @spaarke/auth (peer-dep not yet declared in new package — deferred to a future hoist), ChannelFetchResult
via relative back-pointer to src/solutions/DailyBriefing/src/types/ (types hoist deferred to task 015).

013 (P2 existing hooks hoist): useInlineTodoCreate + useBriefingNarration moved to
@spaarke/daily-briefing-components/src/hooks/. ADR-024 BINDING preservation verified:
- TODO_REGARDING_CATALOG region SHA256 byte-identical to HEAD (101ed595b15f48075df89d046fcaffc11fc2ba2fa11dac3321b3c3a944d64645)
- Full useInlineTodoCreate body SHA256 byte-identical (ec027f2e54bf07acc6646839965790deb80a76fcde8746386f7ff14e27ea792a)
- useBriefingNarration body SHA256 byte-identical (afbb7e1c629469c16ad0472c3af728e80b514cc41bde0bcaab6c3a8fbb4654c2)
Original locations are re-export shims. types/notifications back-pointer matches task 012 precedent.

054 (DD LegalWorkspace + SpaarkeAi auth migration): both solutions' authInit.ts replaced with thin lazy-singleton
factory consumers (same pattern as task 053). Lazy-singleton chosen because both solutions' runtime config getters
are not safe to invoke at module load — setRuntimeConfig fires during bootstrap() AFTER resolveRuntimeConfig
resolves. Direct-destructure would throw "Runtime config not initialized." Config per solution:
- LegalWorkspace: proactiveRefresh true, logLabel 'authInit', tenantId from runtime, 22 consumer imports preserved
- SpaarkeAi: proactiveRefresh true, logLabel 'SpaarkeAi:authInit', tenantId from runtime, 1 consumer import preserved
SpaarkeAi P1 seam preservation evidence: grep for loadSpaarkeAiNotificationContext shows 1 hit (the definition);
0 importers — task 002 deferred per Wave 2a re-sequencing decision. main.tsx files NOT modified by this task.
Per task 053 precedent + user message: FR-20 "delete authInit.ts" interpreted as "eliminate divergent logic" —
files retained as thin wrappers to preserve 23 consumer imports. LW full build green; SAi surface-typecheck green.

032 (P2b prompt + validation unit tests): 4 new tests in
tests/unit/Sprk.Bff.Api.Tests/Api/Ai/DailyBriefingEndpointsTests.cs (EXTENDED existing file per F.2):
- BuildChannelNarrationPrompt_Includes_RegardingId_Per_Item_With_Regarding (FR-15 + FR-16 instruction text)
- BuildChannelNarrationPrompt_Omits_RegardingId_Token_When_Item_Has_No_Regarding (FR-15 graceful-degradation)
- ValidateBulletPrimaryEntityIds_Nulls_Fields_And_Logs_Warning_When_Id_Hallucinated (FR-17 defense in depth)
- ValidateBulletPrimaryEntityIds_Retains_Fields_When_Id_Is_In_Allowed_Set (FR-17 happy path)
All 6 DailyBriefingEndpoints tests pass (2 existing + 4 new). Used MockBehavior.Strict + Mock<IBriefingAi> +
CapturingTestLogger nested class (mirrors AnalysisToolDtoTests canonical pattern). InternalsVisibleTo verified.
Tests directly exercise `internal static` helpers BuildChannelNarrationPrompt, BuildAllowedRegardingIdSet,
ValidateBulletPrimaryEntityIds (per task 031's deliberate `internal` modifier choice).

042 (P3 + P2b combined publish-size + CVE measurement + E2E test plan):
- Compressed publish output: 46.01 MB (48,248,232 bytes); 261 files; uncompressed 139.23 MB
- **Delta vs baseline: +0.00 MB** — well under spec NFR-04 (+1 MB) and §10 ceiling (60 MB)
- CVE diff: 0 new HIGH-severity. Only pre-existing Microsoft.Kiota.Abstractions 1.21.2 (GHSA-7j59-v9qr-6fq9).
- notes/bff-size-p2b-p3.md (combined P2b + P3 report) + notes/p3-e2e-test-plan.md (3 test cases for task 062).
- Task 033 (P2b-only size verification) SUBSUMED by this combined measurement — same code under measurement,
  same baseline reference, identical result. Both marked complete in TASK-INDEX.

Wave 3 ran 6 parallel sub-agents at the concurrency ceiling. Combined post-merge state: 26 BFF tests pass (22 + 4);
36 file changes; new package + LegalWorkspace builds clean; SpaarkeAi surface-typecheck clean (pre-existing
@spaarke/sdap-client Vite error remains — not new). TASK-INDEX.md flipped 011/012/013/032/042/054 to ✅;
033 to ✅ (subsumed); status counter 10 → 17 of 36.
