+++ src/client/shared/Spaarke.DailyBriefing.Components/src/services/index.ts
@@ -3,9 +3,22 @@
  *
  * UI-agnostic data access + AI narration calls. The narration service wraps
  * the BFF `/narrate` endpoint (no new endpoint; uses existing R1 endpoint per
- * spec MUST rule).
+ * spec MUST rule). BFF client lives package-local per Calendar
+ * (`@spaarke/events-components`) precedent — no generic `@spaarke/bff-clients`
+ * package yet.
  *
- * Scaffold placeholder — populated by R2 task 013.
+ * Populated by R2 task 012 (FR-09): hoisted `briefingService` (the BFF
+ * `/summarize` + `/narrate` clients).
  */
 
-export {};
+export {
+  fetchAiBriefing,
+  fetchBriefingNarration,
+  type BriefingResult,
+  type DailyBriefingSummaryResponse,
+  type NarrationResult,
+  type NarrateResponse,
+  type TldrResult,
+  type ChannelNarrationResult,
+  type NarrativeBulletResult,
+} from './briefingService';
