+++ src/solutions/DailyBriefing/src/components/channelIcons.ts
@@ -1,37 +1,16 @@
 /**
- * channelIcons — Maps CHANNEL_REGISTRY iconName strings to Fluent v9 icon
- * components. Centralised here to keep ChannelCard clean.
+ * channelIcons — re-export shim.
  *
- * Only imports from @fluentui/react-icons (ADR-021).
+ * The canonical `getChannelIcon` helper was hoisted alongside the 9 Daily
+ * Briefing components to the new shared package
+ * `@spaarke/daily-briefing-components/components` in R2 task 011 (FR-04).
+ * ActivityNotesSection depends on it so it followed the same hoist.
+ *
+ * Cleanup (replace consumer imports with the canonical package path, then
+ * delete this shim) is tracked by R2 task 017.
+ *
+ * DO NOT add new code here. New work lives in
+ * `src/client/shared/Spaarke.DailyBriefing.Components/src/components/channelIcons.ts`.
  */
 
-import type { FluentIcon } from "@fluentui/react-icons";
-import {
-  WarningRegular,
-  ClockRegular,
-  DocumentRegular,
-  MailRegular,
-  CalendarRegular,
-  BriefcaseRegular,
-  PeopleRegular,
-  InfoRegular,
-} from "@fluentui/react-icons";
-
-const ICON_MAP: Record<string, FluentIcon> = {
-  Warning: WarningRegular,
-  Clock: ClockRegular,
-  Document: DocumentRegular,
-  Mail: MailRegular,
-  Calendar: CalendarRegular,
-  Briefcase: BriefcaseRegular,
-  People: PeopleRegular,
-  Info: InfoRegular,
-};
-
-/**
- * Resolves a CHANNEL_REGISTRY iconName to its Fluent v9 icon component.
- * Falls back to InfoRegular for unknown icon names.
- */
-export function getChannelIcon(iconName: string): FluentIcon {
-  return ICON_MAP[iconName] ?? InfoRegular;
-}
+export { getChannelIcon } from "@spaarke/daily-briefing-components/components";
