+++ src/solutions/DailyBriefing/src/hooks/useInlineTodoCreate.ts
@@ -1,282 +1,17 @@
 /**
- * useInlineTodoCreate — hook for creating `sprk_todo` records directly from
- * notification items in the Daily Briefing.
+ * useInlineTodoCreate — re-export shim.
  *
- * Per smart-todo-decoupling-r3 FR-29 / OS-1 (task 014, 2026-06-08):
- *   - Creates first-class `sprk_todo` Dataverse records.
- *   - Legacy `sprk_event { sprk_todoflag=true, sprk_todostatus, sprk_todosource }`
- *     model is retired across the repo. This hook was the final functional
- *     consumer outside the BFF.
- *   - No backward-compat shim — `sprk_eventtodo` is not written under any
- *     circumstances.
+ * Per R2 task 013 (FR-05): the canonical implementation moved to
+ * `@spaarke/daily-briefing-components/hooks`. This file is preserved as a
+ * thin re-export so existing consumers (App.tsx and any other importers in
+ * `src/solutions/DailyBriefing/src/`) keep building until task 017/018
+ * redirects them to the package import directly.
  *
- * Regarding (multi-entity resolution per ADR-024):
- *   - When the notification item carries `regardingEntityType` + `regardingId`,
- *     the hook resolves the catalog entry from `TODO_REGARDING_CATALOG` and
- *     calls `applyResolverFields` to atomically populate the entity-specific
- *     lookup + four resolver fields (sprk_regardingrecordtype,
- *     sprk_regardingrecordid, sprk_regardingrecordname, sprk_regardingrecordurl).
- *   - If the regarding entity type is not in the catalog, the regarding step is
- *     skipped and the todo is created without an association (the resolver
- *     fields stay null).
- *
- * Tracks per-notification creation state so the UI can show pending/created/error.
- *
- * Usage:
- *   const { createTodo, isCreated, isPending, getError } = useInlineTodoCreate(webApi);
- *
- *   <Button onClick={() => createTodo(notificationItem)} disabled={isPending(item.id)}>
- *     {isCreated(item.id) ? "Created" : "Add To Do"}
- *   </Button>
+ * ADR-024 binding: `TODO_REGARDING_CATALOG` + `applyResolverFields` usage is
+ * preserved byte-identical in the hoisted file.
  */
 
-import { useState, useRef, useCallback } from "react";
-import {
-  applyResolverFields,
-  TODO_REGARDING_CATALOG,
-  type INavPropEntry,
-  type IPolymorphicWebApi,
-} from "@spaarke/ui-components/services";
-import type { IWebApi, NotificationItem, NotificationPriority } from "../types/notifications";
-
-// ---------------------------------------------------------------------------
-// Types
-// ---------------------------------------------------------------------------
-
-type TodoCreateStatus = "pending" | "created" | "error";
-
-export interface UseInlineTodoCreateResult {
-  /** Create a new `sprk_todo` from a notification item. */
-  createTodo: (item: NotificationItem) => Promise<void>;
-  /** Returns true if a To Do was successfully created for this notification. */
-  isCreated: (itemId: string) => boolean;
-  /** Returns true if a To Do creation is in-flight for this notification. */
-  isPending: (itemId: string) => boolean;
-  /** Returns the error message for a failed creation, or undefined. */
-  getError: (itemId: string) => string | undefined;
-}
-
-// ---------------------------------------------------------------------------
-// sprk_todo lifecycle constants (mirrors SmartTodo/DataverseService.ts task 020)
-// ---------------------------------------------------------------------------
-
-/** sprk_todo statecode: 0 = Active. */
-const STATECODE_ACTIVE = 0;
-/** sprk_todo statuscode: 1 = Open (statecode 0). */
-const STATUSCODE_OPEN = 1;
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-/**
- * Map notification priority string to a 0-100 priority score for
- * `sprk_todo.sprk_priorityscore`. Bands match the implicit kanban
- * banding used elsewhere (low ~20, normal ~50, high ~70, urgent ~90).
- */
-function mapPriorityToScore(priority: NotificationPriority): number {
-  switch (priority) {
-    case "urgent":
-      return 90;
-    case "high":
-      return 70;
-    case "normal":
-      return 50;
-    case "low":
-      return 20;
-    default:
-      return 50;
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Nav-prop discovery for sprk_todo (mirrors the shared `todoService.ts` pattern)
-// ---------------------------------------------------------------------------
-
-/** Page-session cache of discovered ManyToOne nav-props, keyed by entity name. */
-const _navPropCache: Record<string, INavPropEntry[]> = {};
-
-async function _discoverNavProps(
-  entityLogicalName: string
-): Promise<INavPropEntry[]> {
-  if (_navPropCache[entityLogicalName]) {
-    return _navPropCache[entityLogicalName];
-  }
-
-  try {
-    const url =
-      `/api/data/v9.0/EntityDefinitions(LogicalName='${entityLogicalName}')/ManyToOneRelationships` +
-      `?$select=ReferencingAttribute,ReferencingEntityNavigationPropertyName,ReferencedEntity`;
-
-    const resp = await fetch(url, { credentials: "include" });
-    if (!resp.ok) {
-      console.warn(
-        `[useInlineTodoCreate] Nav-prop discovery failed for ${entityLogicalName}: HTTP ${resp.status}`
-      );
-      return [];
-    }
-
-    const json = (await resp.json()) as {
-      value?: Array<{
-        ReferencingAttribute: string;
-        ReferencingEntityNavigationPropertyName: string;
-        ReferencedEntity: string;
-      }>;
-    };
-
-    const entries: INavPropEntry[] = (json.value ?? []).map((r) => ({
-      columnName: r.ReferencingAttribute,
-      navPropName: r.ReferencingEntityNavigationPropertyName,
-      referencedEntity: r.ReferencedEntity,
-    }));
-
-    _navPropCache[entityLogicalName] = entries;
-    return entries;
-  } catch (err) {
-    console.warn(
-      `[useInlineTodoCreate] Nav-prop discovery error for ${entityLogicalName}:`,
-      err
-    );
-    return [];
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Hook
-// ---------------------------------------------------------------------------
-
-/**
- * Hook for inline To Do creation from notification items.
- *
- * @param webApi - Xrm.WebApi reference (from xrmProvider). Pass null if not yet available.
- * @returns Object with createTodo, isCreated, isPending, getError functions.
- */
-export function useInlineTodoCreate(webApi: IWebApi | null): UseInlineTodoCreateResult {
-  const [statusMap, setStatusMap] = useState<Map<string, TodoCreateStatus>>(
-    () => new Map()
-  );
-  const errorsRef = useRef<Map<string, string>>(new Map());
-
-  const createTodo = useCallback(
-    async (item: NotificationItem): Promise<void> => {
-      if (!webApi) {
-        console.error("[useInlineTodoCreate] webApi not available");
-        return;
-      }
-
-      // Optimistic: set to pending
-      setStatusMap((prev) => {
-        const next = new Map(prev);
-        next.set(item.id, "pending");
-        return next;
-      });
-
-      try {
-        // 1. Build the core sprk_todo record (per entity-schema.md + SmartTodo
-        //    DataverseService.createTodo pattern).
-        const record: Record<string, unknown> = {
-          sprk_name: item.title,
-          statecode: STATECODE_ACTIVE,
-          statuscode: STATUSCODE_OPEN,
-          sprk_priorityscore: mapPriorityToScore(item.priority),
-        };
-
-        if (item.body) {
-          // sprk_notes is the rich-text/long-text body field on sprk_todo
-          // (replaces the legacy sprk_event.sprk_description path).
-          record["sprk_notes"] = item.body;
-        }
-
-        // 2. Apply regarding (multi-entity resolution per ADR-024) — only when
-        //    a regarding entity is supplied AND it's in the supported catalog.
-        if (item.regardingEntityType && item.regardingId) {
-          const catalogEntry = TODO_REGARDING_CATALOG.find(
-            (c: { entityType: string }) => c.entityType === item.regardingEntityType
-          );
-
-          if (catalogEntry) {
-            // Wrap IWebApi.retrieveMultipleRecords into the shape
-            // applyResolverFields expects (IPolymorphicWebApi).
-            const polyWebApi: IPolymorphicWebApi = {
-              retrieveMultipleRecords: (entityLogicalName: string, query: string) =>
-                webApi.retrieveMultipleRecords(entityLogicalName, query),
-            };
-
-            const navProps = await _discoverNavProps("sprk_todo");
-
-            // Notification carries the parent's display name in `regardingName`.
-            await applyResolverFields(
-              polyWebApi,
-              record,
-              navProps,
-              catalogEntry.entityType,
-              catalogEntry.entitySet,
-              item.regardingId,
-              item.regardingName ?? "",
-              catalogEntry.navPropHint
-            );
-          } else {
-            // Regarding entity not in supported catalog — create the todo
-            // without a regarding association rather than failing the whole
-            // operation. This handles notifications for entities not in the
-            // 11-entity multi-entity resolution scope.
-            console.info(
-              `[useInlineTodoCreate] Regarding entity "${item.regardingEntityType}" not in TODO_REGARDING_CATALOG — creating todo without association.`
-            );
-          }
-        }
-
-        // 3. Create the sprk_todo record (NEVER sprk_event).
-        await webApi.createRecord("sprk_todo", record);
-
-        // Success
-        setStatusMap((prev) => {
-          const next = new Map(prev);
-          next.set(item.id, "created");
-          return next;
-        });
-      } catch (e: unknown) {
-        // Failure
-        let message = "Failed to create To Do. Please try again.";
-        if (e && typeof e === "object") {
-          const err = e as Record<string, unknown>;
-          if (typeof err["message"] === "string") {
-            message = err["message"];
-          }
-        } else if (typeof e === "string") {
-          message = e;
-        }
-
-        errorsRef.current.set(item.id, message);
-        console.error("[useInlineTodoCreate] Failed to create To Do:", message, e);
-
-        setStatusMap((prev) => {
-          const next = new Map(prev);
-          next.set(item.id, "error");
-          return next;
-        });
-      }
-    },
-    [webApi]
-  );
-
-  const isCreated = useCallback(
-    (itemId: string): boolean => statusMap.get(itemId) === "created",
-    [statusMap]
-  );
-
-  const isPending = useCallback(
-    (itemId: string): boolean => statusMap.get(itemId) === "pending",
-    [statusMap]
-  );
-
-  const getError = useCallback(
-    (itemId: string): string | undefined => errorsRef.current.get(itemId),
-    // statusMap is included so the callback identity changes when errors are recorded
-    // (error status is set in statusMap at the same time as errorsRef is updated)
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [statusMap]
-  );
-
-  return { createTodo, isCreated, isPending, getError };
-}
+export {
+  useInlineTodoCreate,
+  type UseInlineTodoCreateResult,
+} from "@spaarke/daily-briefing-components/hooks";
